Change "$mount == undef" to "$share != undef"
Because $mount gets set to the namevar, it can never be undef and so the share parameter will never be used.  When used as specified in the README, the client mount exported from the server export definition always gets a namevar with spaces, and tries to use that as the mount name, yielding an error like this:

Error: Failed to apply catalog: Parameter name failed on Mount[shared /opt/sge by compute1.example.com on shared /opt/sge by head.example.com]: name must not contain whitespace: shared /opt/sge by head.example.com at /etc/puppet/environments/hpc/modules/nfs/manifests/client/mount.pp:115

Since nfs::server::export declares the exported Nfs::Client::Mount resource using the 'share' parameter instead of the 'mount' parameter, we need to look at whether or not share is undef; if it is, ignore it, otherwise, use it as the mount.

I have not tested this for compatibility with NFSv4 or whatever else in the module that might break; it got my setup (and the examples in the README) to work and I'm leaving it at that.